requirements.txt
python-telegram-bot==21.3
python-gitlab==4.7.0
requests~=2.32.3
python-dotenv~=1.0.1
style~=1.1.0
pip~=24.0
attrs~=23.2.0



